backend/pyproject.toml
[project]
name = "athena-eye-backend"
version = "0.1.0"
description = "The backend API and worker for the Athena Eye project."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.6",
    "fastapi>=0.116.1",
    "gunicorn>=23.0.0",
    "mysql-connector-python>=9.4.0",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.42",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.35.0",
]


[tool.setuptools]

packages = ["athena_eye_project"]
